[InfoToolTips] Convert to TypeScript
