--- services/frontend/src/common/InfoToolTips/closeToGraduation.js
@@ -1,14 +1,0 @@
-export const closeToGraduationToolTips = `
-  You can filter students by choosing a degree programme and/or a faculty. You can also select multiple programmes or faculties. Please note that if you select a faculty, only the degree programmes of that faculty will be shown.  
-
-  Students are included in this list based on the number of credits they have completed in their primary study plan. The limits for each study programme are as follows:  
-
-  - All bachelor’s programmes: 140 credits
-  - Degree Programme in Veterinary Medicine (eläinlääketieteen lisensiaatin koulutusohjelma): 150 credits
-  - Degree Programme in Medicine (lääketieteen koulutusohjelma): 330 credits
-  - Degree Programme in Dentistry (hammaslääketieteen koulutusohjelma): 300 credits
-  - Master’s Programme in Psychology (psykologian maisteriohjelma): 115 credits
-  - Other master’s programmes: 70 credits
-
-  If you have access to all student data in Oodikone, you can also see all students here. Otherwise you will only see the students in your study guidance groups.
-  `
