+++ services/frontend/src/common/InfoToolTips/courseStatistics.ts
