+++ services/frontend/src/common/InfoToolTips/index.ts
@@ -1,9 +1,9 @@
-export { courseStatisticsToolTips } from './courseStatistics'
+export { closeToGraduationToolTips } from './closeToGraduation'
 export { coursePopulationToolTips } from './coursePopulation'
+export { courseStatisticsToolTips } from './courseStatistics'
 export { facultyToolTips } from './faculty'
 export { filterToolTips } from './filters'
 export { populationStatisticsToolTips } from './populationStatistics'
 export { studentToolTips } from './students'
-export { studyProgrammeToolTips } from './studyprogramme'
+export { studyProgrammeToolTips } from './studyProgramme'
 export { userToolTips } from './users'
-export { closeToGraduationToolTips } from './closeToGraduation'
