--- services/frontend/src/common/InfoToolTips/students.js
@@ -1,10 +1,0 @@
-export const studentToolTips = {
-  StudyrightStatus: `
-  Shows the status of the current semester enrollment of the student in the corresponding programme.
-
-  **Active:** Student has a semester enrollment to the corresponding programme as either present or absent.  
-  **Inactive:** Student has not enrolled to the corresponding programme this semester.  
-  **Cancelled:** Studyright for the corresponding programme is cancelled.  
-  **Graduated:** Student has graduated from the corresponding programme.
-  `,
-}
