+++ services/frontend/src/common/InfoToolTips/students.ts
@@ -1,0 +1,10 @@
+export const studentToolTips = {
+  studyRightStatus: `
+    Shows the status of the current semester enrollment of the student in the corresponding programme.
+
+    **Active:** Student has a semester enrollment to the corresponding programme as either present or absent.  
+    **Inactive:** Student has not enrolled to the corresponding programme this semester.  
+    **Cancelled:** Study right for the corresponding programme is cancelled.  
+    **Graduated:** Student has graduated from the corresponding programme.
+  `,
+}
